backend/pyproject.toml
[project]
name = "disaster-resource-management"
version = "1.0.0"
description = "AI-powered disaster monitoring, prediction, victim assistance and resource allocation platform"
requires-python = ">=3.11,<3.14"
license = "MIT"
dependencies = [
    # Web framework
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "python-dotenv>=1.0,<2.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.1,<3.0",
    # Database (Supabase PostgREST client)
    "supabase>=2.0,<3.0",
    # ML / Data Science
    "scikit-learn>=1.4,<1.7",
    "imbalanced-learn>=0.12,<0.14",
    "xgboost>=2.0,<3.0",
    "pandas>=2.2,<3.0",
    "numpy>=1.26,<3.0",
    "joblib>=1.3,<2.0",
    # Auth & Security
    "python-jose[cryptography]>=3.3,<4.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "python-multipart>=0.0.9,<1.0",
    # Async / HTTP
    "aiofiles>=23.2,<25.0",
    "httpx>=0.25,<1.0",
    # Caching & Task Queue
    "redis>=5.0,<6.0",
    "celery>=5.3,<6.0",
    # Rate Limiting
    "slowapi>=0.1.9,<1.0",
    # Forecasting & Optimization
    "prophet>=1.1,<2.0",
    "PuLP>=2.7,<4.0",
    "ortools>=9.8,<10.0",
    # Phase 4: Real-Time Data Ingestion
    "feedparser>=6.0,<7.0",
    # Phase 5: AI Operations – email delivery
    "sendgrid>=6.11,<7.0",
    # Phase 5: AI Operations – scheduling + reporting
    "apscheduler>=3.10,<4.0",
    "alembic>=1.18.4",
    "plotly>=6.5.2",
    # Phase: TFT Severity Forecasting
    "torch>=2.0,<3.0",
    "lightning>=2.0,<3.0",
    "pytorch-forecasting>=1.0,<2.0",
    # Phase: GAT Resource Allocator
    "torch-geometric>=2.5,<3.0",
    # Phase: Causal AI (DoWhy + PDF reports)
    "dowhy>=0.11,<1.0",
    "reportlab>=4.0,<5.0",
    # Phase: DisasterGPT — RAG + LLM
    "chromadb>=0.5,<1.0",
    "sentence-transformers>=3.0,<4.0",
    "beautifulsoup4>=4.12,<5.0",
    "llama-cpp-python>=0.3,<1.0",
    "groq>=0.11,<1.0",
    "asyncpg>=0.31.0",
]

[project.optional-dependencies]
finetune = [
    # Only needed on the training machine (requires GPU)
    "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git",
    "trl>=0.8,<1.0",
    "peft>=0.10,<1.0",
    "accelerate>=0.28,<1.0",
    "bitsandbytes>=0.43,<1.0",
    "datasets>=3.4.1,<4.4.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=4.1,<6.0",
    "ruff>=0.4,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=4.1,<6.0",
    "ruff>=0.4,<1.0",
]

[tool.setuptools.packages.find]
include = ["app*", "ml*", "scripts*"]
exclude = ["tests*", "logs*", "models*", "training_data*", "lightning_logs*", "migrations*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
